% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Dire bonjour}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{Une chaine de carateres contenant le prenom de la personne à qui dire bonjour}
}
\value{
Une chaine de caracteres
}
\description{
Dire bonjour à la personne dont le prénom est donné
}
\examples{
library(clementclaustre)
#Comportement par défaut
dire_bonjour()
#Dire bonjour à Clément
dire_bonjour("Clément")

}
