% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\docType{data}
\name{pays}
\alias{pays}
\title{pays}
\format{
data frame de 241 lignes et 6 colonnes
\describe{
\item{ V1 }{  int }
\item{ V2 }{  int }
\item{ V3 }{  chr }
\item{ V4 }{  chr }
\item{ V5 }{  chr }
\item{ V6 }{  chr }
}
}
\usage{
pays
}
\description{
Un jeu de données contenant des informations sur des pays
}
\keyword{datasets}
